- Fixed bug that was preventing the model specification and the specchangelog from being saved.  - Added the ability to go to a specific class page of a manufacturer from the classes page.

+++ TractorSpecs.Core/Models/SpecChangeLog.cs
@@ -11,7 +11,7 @@ namespace TractorSpecs.Core.Models
     {
         public long? modelId { get; set; }
 
-        public long? specificationId { get; set; }
+        public int? specificationId { get; set; }
 
         [StringLength(50)]
         public string specValue { get; set; }
@@ -22,7 +22,7 @@ namespace TractorSpecs.Core.Models
         [StringLength(50)]
         public string source { get; set; }
 
-        public long specChangeLogId { get; set; }
+        public int specChangeLogId { get; set; }
 
         public virtual model model { get; set; }
